Initial Commit. See notes.
Rearranged many functions in order to get this to work right. Improvements in main functionality are there, but there are flaws with advisorcounting.

First: the counts seems to be incrementing right; this is done using a lambda function and the mapRows method on CSVObject. But the values aren't printing where they should in the exported files.

Second: We should dedup somehow when incrementing. Maybe throw in a list of student ID's on a class, and use it to dedup? We can use that for other features as well, so keep this in mind when implementing.

Minor: changed variable names. This might make trouble when merging to main later.